bench: keep the V harness's ids int32, matching the Go side
The cross-language byte-compat check went red: the Go harness builds ids
with int32(r.next() & 0xffffffff), and V's int() matched that only while
`int` was 32-bit. Since vlang/v#28293 it is 64-bit, so V kept the ids
positive where Go wrapped them negative and the two encodings diverged
at the first id.

Byte-identical again in both directions, and the perf gate passes.

+++ bench/v/main.v
@@ -29,7 +29,9 @@ fn build_book(n int) AddressBook {
 	mut people := []Person{cap: n}
 	for i in 0 .. n {
 		pad := int(r.next() % 24)
-		id := int(r.next() & 0xffffffff)
+		// i32 to match the Go harness's int32(): `int` is 64-bit since
+		// vlang/v#28293, so it no longer truncates these into negative ids
+		id := int(i32(r.next() & 0xffffffff))
 		active := (r.next() & 1) == 1
 		score := f64(r.next() % 100000) / 100.0
 		last_seen := r.next()
